// src/app/models/cuotaMortuaria/datos-ejecutivo.model.ts
export class DatosEjecutivo {

    IDEJECUTIVO: string;

}
export class DatosEjecutivoResponse {

    CODIGOCARGO: string;
    CARGO: string;
    CODIGOCENTROCOSTO: string;
    CENTROCOSTO: string;
    CODIGOSUCURSAL: string;
    SUCURSAL: string;
    FECHAINGRESO: string;
    FECHAFINIQUITO: string;
    CODIGOMARCA: string;
    NOMBRES: string;
    APELLIDOPATERNO: string;
    APELLIDOMATERNO: string;
    RUTEMPLEADO: string;
    DIGITORUTEMPLEADO: string;
    EMAIL: string;
    RUTJEFATURA: string;
    DIGITORUTJEFATURA: string;
    USUARIODATABASE: string;
    USUARIONT: string;
}