// src/app/models/cuotaMortuaria/poliza.model.ts
export class Poliza {
    P_POLIZA: number;
}

export class PolizaResponse {

    POL_LINEA: number;
    POL_PRODUCTO: number;
    POL_DOCUMENTO: number;
    POL_POLIZA: number;
    POL_ATRB07: number;
    POL_ATRB06: number;
    NUM_RUT_NAT: number;
    NAT_ID: number;
    FEC_INICIO_PAGO: string;
    SIN_TIPO: number;
}